# docs/swagger.yaml
basePath: /api
definitions:
  handler.errorResponse:
    properties:
      message:
        type: string
    type: object
...
  model.Material:
    properties:
      coefficient:
        type: number
      description:
        type: string
      id:
        type: integer
      image_url:
        type: string
      is_deleted:
        type: boolean
      title:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API для расчета земляных работ при разработке котлована
  title: DIA API
  version: "1.0"
paths:
  /calculation-materials/{calculation_id}/{material_id}:
    delete:
      consumes:
      - application/json
      description: Удаление материала из заявки расчета (только владелец заявки)
      parameters:
      - description: ID заявки расчета
        in: path
        name: calculation_id
        required: true
        type: integer
      - description: ID материала
        in: path
        name: material_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - BearerAuth: []
      summary: Удаление материала из заявки
      tags:
      - calculation-materials
    put:
      consumes:
      - application/json
      description: Обновление угла откоса материала в заявке расчета (только владелец
        заявки)
      parameters:
      - description: ID заявки расчета
        in: path
        name: calculation_id
        required: true
        type: integer
      - description: ID материала
        in: path
        name: material_id
        required: true
        type: integer
      - description: Угол откоса
        in: body
        name: input
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - BearerAuth: []
      summary: Обновление материала в заявке
      tags:
      - calculation-materials
  /materials:
    get:
      consumes:
      - application/json
      description: Получение списка всех материалов с возможностью фильтрации по названию
      parameters:
      - description: Фильтр по названию материала
        in: query
        name: materialTitle
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Material'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: Получение списка материалов
      tags:
      - materials
    post:
      consumes:
      - application/json
      description: Создание материала (только для модераторов)
      parameters:
      - description: Данные материала
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/model.Material'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Material'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - BearerAuth: []
      summary: Создание нового материала
      tags:
      - materials
  /materials/{id}:
    delete:
      consumes:
      - application/json
      description: Удаление материала (только для модераторов)
      parameters:
      - description: ID материала
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - BearerAuth: []
      summary: Удаление материала
      tags:
      - materials
    get:
      consumes:
      - application/json
      description: Получение информации о конкретном материале
      parameters:
      - description: ID материала
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Material'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: Получение материала по ID
      tags:
      - materials
    put:
      consumes:
      - application/json
      description: Обновление информации о материале (только для модераторов)
      parameters:
      - description: ID материала
        in: path
        name: id
        required: true
        type: integer
      - description: Данные для обновления
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/model.Material'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Material'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - BearerAuth: []
      summary: Обновление материала
      tags:
      - materials
  /materials/{id}/image:
    post:
      consumes:
      - multipart/form-data
      description: Загрузка изображения для материала (только для модераторов)
      parameters:
      - description: ID материала
        in: path
        name: id
        required: true
        type: integer
      - description: Изображение материала
        in: formData
        name: image
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - BearerAuth: []
      summary: Загрузка изображения материала
      tags:
      - materials
  /materials/{id}/pit:
    post:
      consumes:
      - application/json
      description: Добавление материала в текущую заявку пользователя
      parameters:
      - description: ID материала
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "201":
          description: Created
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - BearerAuth: []
      summary: Добавление материала в заявку
      tags:
      - materials
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    description: 'Введите JWT токен в формате: "Bearer {your_token}".'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"